refactor(dat): extract shared helpers and RowCtx from DatComposer (#611 PR1)
Extract RowCtx (renamed DatRowContext), CollectionMetadataValues, and
static helpers (ResolveHashColumn, GetEffectiveProfileGenerator) to
separate files. Pure file move, no behavior change. DatComposer.cs
728 to 653 lines. Forward-looking seam for PR2-4 mode composer extraction.
