+++ src/LoadFiles/CollectionMetadataValues.cs
@@ -1,0 +1,7 @@
+namespace Zipper.LoadFiles;
+
+internal static class CollectionMetadataValues
+{
+    public static readonly string[] DataSources = ["Email Server", "Network Share", "Cloud Storage", "Mobile Device", "Laptop", "Desktop", "Server", "External Media"];
+    public static readonly string[] ProcessingStatuses = ["Processed", "Pending Review", "Approved", "Privileged", "Redacted", "Produced"];
+}
